TriggerBuffs : Fix Poison buff display
Fix message sent to client when it's a DirectDamage with duration (poison). It now display "11 to 12 damage air" in the buff instead of crap.

I need more experience to understand why only this effet is not displayed properly. I will update it IF or WHEN I find another issue like this one.